+++ internal/impl/sql/conn_fields.go
@@ -1,0 +1,70 @@
+package sql
+
+import (
+	"database/sql"
+	"time"
+
+	"github.com/benthosdev/benthos/v4/public/service"
+)
+
+func connFields() []*service.ConfigField {
+	return []*service.ConfigField{
+		service.NewDurationField("conn_max_idle_time").
+			Description(`An optional maximum amount of time a connection may be idle. Expired connections may be closed lazily before reuse. If value <= 0, connections are not closed due to a connection's idle time.`).
+			Optional().
+			Advanced(),
+		service.NewDurationField("conn_max_lifetime").
+			Description(`An optional maximum amount of time a connection may be reused. Expired connections may be closed lazily before reuse. If value <= 0, connections are not closed due to a connection's age.`).
+			Optional().
+			Advanced(),
+		service.NewIntField("conn_max_idle").
+			Description(`An optional maximum number of connections in the idle connection pool. If conn_max_open is greater than 0 but less than the new conn_max_idle, then the new conn_max_idle will be reduced to match the conn_max_open limit. If value <= 0, no idle connections are retained. The default max idle connections is currently 2. This may change in a future release.`).
+			Optional().
+			Advanced(),
+		service.NewIntField("conn_max_open").
+			Description(`An optional maximum number of open connections to the database. If conn_max_idle is greater than 0 and the new conn_max_open is less than conn_max_idle, then conn_max_idle will be reduced to match the new conn_max_open limit. If value <= 0, then there is no limit on the number of open connections. The default is 0 (unlimited).`).
+			Optional().
+			Advanced(),
+	}
+}
+
+type connSettings struct {
+	connMaxLifetime time.Duration
+	connMaxIdleTime time.Duration
+	maxIdleConns    int
+	maxOpenConns    int
+}
+
+func (c connSettings) apply(db *sql.DB) {
+	db.SetConnMaxIdleTime(c.connMaxIdleTime)
+	db.SetConnMaxLifetime(c.connMaxLifetime)
+	db.SetMaxIdleConns(c.maxIdleConns)
+	db.SetMaxOpenConns(c.maxOpenConns)
+}
+
+func connSettingsFromParsed(conf *service.ParsedConfig) (c connSettings, err error) {
+	if conf.Contains("conn_max_lifetime") {
+		if c.connMaxLifetime, err = conf.FieldDuration(); err != nil {
+			return
+		}
+	}
+
+	if conf.Contains("conn_max_idle_time") {
+		if c.connMaxIdleTime, err = conf.FieldDuration("conn_max_idle_time"); err != nil {
+			return
+		}
+	}
+
+	if conf.Contains("conn_max_idle") {
+		if c.maxIdleConns, err = conf.FieldInt("conn_max_idle"); err != nil {
+			return
+		}
+	}
+
+	if conf.Contains("conn_max_open") {
+		if c.maxOpenConns, err = conf.FieldInt("conn_max_open"); err != nil {
+			return
+		}
+	}
+	return
+}