+++ internal/impl/sql/input_sql_select.go
@@ -5,7 +5,6 @@ import (
 	"database/sql"
 	"fmt"
 	"sync"
-	"time"
 
 	"github.com/Masterminds/squirrel"
 
@@ -15,7 +14,7 @@ import (
 )
 
 func sqlSelectInputConfig() *service.ConfigSpec {
-	return service.NewConfigSpec().
+	spec := service.NewConfigSpec().
 		// Stable(). TODO
 		Categories("Services").
 		Summary("Executes a select query and creates a message for each row received.").
@@ -45,32 +44,13 @@ func sqlSelectInputConfig() *service.ConfigSpec {
 		Field(service.NewStringField("suffix").
 			Description("An optional suffix to append to the select query.").
 			Optional().
-			Advanced()).
-		Field(service.NewDurationField("conn_max_idle_time").
-			Description(`An optional maximum amount of time a connection may be idle.
-Expired connections may be closed lazily before reuse.
-If value <= 0, connections are not closed due to a connection's idle time.`).
-			Optional().
-			Advanced()).
-		Field(service.NewDurationField("conn_max_lifetime").
-			Description(`An optiona maximum amount of time a connection may be reused.
-Expired connections may be closed lazily before reuse.
-If value <= 0, connections are not closed due to a connection's age.`).
-			Optional().
-			Advanced()).
-		Field(service.NewIntField("max_idle_cons").
-			Description(`An optional maximum number of connections in the idle connection pool.
-If MaxOpenConns is greater than 0 but less than the new MaxIdleConns, then the new MaxIdleConns will be reduced to match the MaxOpenConns limit.
-If value <= 0, no idle connections are retained.
-The default max idle connections is currently 2. This may change in a future release.`).
-			Optional().
-			Advanced()).
-		Field(service.NewIntField("max_open_conns").
-			Description(`An optional maximum number of open connections to the database.
-If MaxIdleConns is greater than 0 and the new MaxOpenConns is less than MaxIdleConns, then MaxIdleConns will be reduced to match the new MaxOpenConns limit.
-If value <= 0, then there is no limit on the number of open connections. The default is 0 (unlimited).`).
-			Optional().
-			Advanced()).
+			Advanced())
+
+	for _, f := range connFields() {
+		spec = spec.Field(f)
+	}
+
+	spec = spec.
 		Version("3.59.0").
 		Example("Consume a Table (PostgreSQL)",
 			`
@@ -89,6 +69,7 @@ input:
       ]
 `,
 		)
+	return spec
 }
 
 func init() {
@@ -120,10 +101,7 @@ type sqlSelectInput struct {
 	where       string
 	argsMapping *bloblang.Executor
 
-	connMaxLifetime time.Duration
-	connMaxIdleTime time.Duration
-	maxIdleConns    int
-	maxOpenConns    int
+	connSettings connSettings
 
 	logger  *service.Logger
 	shutSig *shutdown.Signaller
@@ -188,40 +166,9 @@ func newSQLSelectInputFromConfig(conf *service.ParsedConfig, logger *service.Log
 		s.builder = s.builder.Suffix(suffixStr)
 	}
 
-	if conf.Contains("conn_max_lifetime") {
-		connMaxLifetime, err := conf.FieldDuration()
-		if err != nil {
-			return nil, err
-		}
-		s.connMaxLifetime = connMaxLifetime
+	if s.connSettings, err = connSettingsFromParsed(conf); err != nil {
+		return nil, err
 	}
-
-	if conf.Contains("conn_max_idle_time") {
-		connMaxIdleTime, err := conf.FieldDuration("conn_max_idle_time")
-		if err != nil {
-			return nil, err
-		}
-		s.connMaxIdleTime = connMaxIdleTime
-	}
-
-	if conf.Contains("max_idle_conns") {
-		maxIdleCons, err := conf.FieldInt("max_idle_conns")
-		if err != nil {
-			return nil, err
-		}
-		s.maxIdleConns = maxIdleCons
-	} else {
-		s.maxIdleConns = 2 // default value for the library
-	}
-
-	if conf.Contains("max_open_conns") {
-		maxOpenConns, err := conf.FieldInt("max_open_conns")
-		if err != nil {
-			return nil, err
-		}
-		s.maxOpenConns = maxOpenConns
-	}
-
 	return s, nil
 }
 
@@ -243,10 +190,7 @@ func (s *sqlSelectInput) Connect(ctx context.Context) (err error) {
 		}
 	}()
 
-	db.SetConnMaxIdleTime(s.connMaxIdleTime)
-	db.SetConnMaxLifetime(s.connMaxLifetime)
-	db.SetMaxIdleConns(s.maxIdleConns)
-	db.SetMaxOpenConns(s.maxOpenConns)
+	s.connSettings.apply(db)
 
 	var args []interface{}
 	if s.argsMapping != nil {